split up mlonmcu/setup/tasks.py into several files
